+++ frontend/src/pages/RAGDocuments/components/QueryArea.tsx
@@ -87,6 +87,7 @@ function QueryArea() {
           `doc_tasks/${doc_task_id}`
         );
         const data = response.data;
+        console.log(data);
         current.setFromServerResponse(data);
         docTasksShortStore.setNeedReload(true);
         // Continue polling until status becomes 6
@@ -178,6 +179,21 @@ function QueryArea() {
     fetchDocTaskOptions();
   }, [docTaskOptionsStore.needReload]);
   useEffect(() => docTaskOptionsStore.setNeedReload(true), []);
+
+  const tokensCounterVisible =
+    (current.vdb_query_seconds ||
+      current.llm_query_seconds ||
+      current.llm_tokens_sent ||
+      current.llm_tokens_received) !== null;
+  function secondsOrNA(seconds: number | null) {
+    if (seconds === null) return "N/A";
+    return `${seconds} sec`;
+  }
+  const tokensCounterText = `VectorDB/LLM time: ${secondsOrNA(
+    current.vdb_query_seconds
+  )} / ${secondsOrNA(current.llm_query_seconds)} Tokens Sent/Recv: ${
+    current.llm_tokens_sent || "N/A"
+  }/${current.llm_tokens_received || "N/A"}`;
 
   return (
     <div className="p-3 border-bottom bg-light">
@@ -293,6 +309,12 @@ function QueryArea() {
         <ProgressBar animated now={current.status_pct || 0} />
       </div>
       <h5>{current.status_text || ""}</h5>
+      <small
+        className="m-0 fw-bold"
+        style={{ visibility: tokensCounterVisible ? "visible" : "hidden" }}
+      >
+        {tokensCounterText}
+      </small>
     </div>
   );
 }